refactor functions into multiple files
