+++ ryanwallace.cloud/map/src/map.ts
@@ -6,94 +6,25 @@ import 'leaflet.fullscreen'
 import 'leaflet-easybutton'
 import 'leaflet-arrowheads'
 import 'invert-color'
-import DOMPurify from 'dompurify'
-import invert from 'invert-color'
-import DataTable from 'datatables.net-dt'
 import { MaptilerLayer } from '@maptiler/leaflet-maptilersdk'
-import { formatDistance } from 'date-fns'
 
-// Cookie utility functions
-function setCookie(name: string, value: string, days: number = 365): void {
-  const expires = new Date()
-  expires.setTime(expires.getTime() + days * 24 * 60 * 60 * 1000)
-  document.cookie = `${name}=${value};expires=${expires.toUTCString()};path=/`
-}
-
-function getCookie(name: string): string | null {
-  const nameEQ = name + '='
-  const ca = document.cookie.split(';')
-  for (let i = 0; i < ca.length; i++) {
-    let c = ca[i]
-    while (c.charAt(0) === ' ') c = c.substring(1, c.length)
-    if (c.indexOf(nameEQ) === 0) return c.substring(nameEQ.length, c.length)
-  }
-  return null
-}
-
-// Type definitions
-interface VehicleFeature {
-  id?: string | number
-  geometry: {
-    type: 'Point' | 'LineString'
-    coordinates: number[]
-  }
-  properties: {
-    'marker-symbol'?: string
-    'marker-color'?: string
-    'marker-size'?: string
-    route?: string
-    status?: string
-    stop?: string
-    update_time?: string
-    speed?: number
-    approximate_speed?: boolean
-    occupancy_status?: string
-    carriages?: string[]
-    stop_eta?: string
-    'stop-coordinates'?: number[]
-    name?: string
-    headsign?: string
-    platform_prediction?: string
-  }
-}
-
-interface AlertEntity {
-  attributes: {
-    header: string
-    severity: string
-    updated_at?: string
-    created_at: string
-    active_period: Array<{
-      end?: string
-    }>
-    informed_entity: Array<{
-      route: string
-    }>
-  }
-}
-
-interface AlertData {
-  data: AlertEntity[]
-}
-
-interface RouteMapping {
-  svg: string
-  alt: string
-}
-
-declare global {
-  interface Window {
-    $: typeof import('jquery')
-  }
-}
+// Import modules
+import { setCookie, getCookie, return_colors } from './utils'
+import { pointToLayer, onEachFeature } from './markers'
+import {
+  layerGroups,
+  shapesLayerGroups,
+  getShapesLayerGroupForRoute
+} from './layer-groups'
+import { updateMarkers } from './marker-manager'
+import { updateTable } from './table-manager'
+import { alerts } from './alerts'
+import { fetchAmtrakData } from './amtrak'
 
 // Extend jQuery to include getJSON method
 declare const $: {
   getJSON: (url: string, callback: (data: any) => void) => void
 }
-
-const markerIdInfo: Map<string, { route: string; vehicleType: string }> =
-  new Map()
 
 var map = L.map('map', {
   doubleTouchDragZoom: true,
@@ -106,17 +37,12 @@ var map = L.map('map', {
   }
 }).setView([42.36565, -71.05236], 13)
 
-const lines: string[] = ['rl', 'gl', 'bl', 'ol', 'sl', 'cr', 'amtrak']
-const vehicleTypes: string[] = ['light', 'heavy', 'regional', 'bus']
-const vehicleCountMap: Map<
-  string,
-  Map<string, number>
-> = createVehicleCountMap()
 const vehicles_url: string =
   process.env.VEHICLES_URL || 'https://vehicles.ryanwallace.cloud'
 const bos_url: string = 'https://bos.ryanwallace.cloud'
 
 let baseLayerLoaded: boolean = false
+let buildingMarkers: L.GeoJSON | null = null
 
 document.getElementById('map')?.scrollIntoView({ behavior: 'smooth' })
 
@@ -124,658 +50,6 @@ new MaptilerLayer({
   apiKey: process.env.MT_KEY || '',
   style: 'streets-v2'
 }).addTo(map)
-
-let currentMarkers: Map<string | number, L.Marker> = new Map()
-let buildingMarkers: L.GeoJSON | null = null
-
-const layerGroups = {
-  red: L.layerGroup(),
-  blue: L.layerGroup(),
-  green: L.layerGroup(),
-  orange: L.layerGroup(),
-  silver: L.layerGroup(),
-  commuter: L.layerGroup(),
-  amtrak: L.layerGroup()
-}
-
-const shapesLayerGroups = {
-  red: L.layerGroup(),
-  blue: L.layerGroup(),
-  green: L.layerGroup(),
-  orange: L.layerGroup(),
-  silver: L.layerGroup(),
-  commuter: L.layerGroup()
-}
-
-function return_colors(route: string): string {
-  if (route.startsWith('Green')) {
-    return '#008150'
-  }
-  if (route.startsWith('Blue')) {
-    return '#2F5DA6'
-  }
-  if (route.startsWith('CR')) {
-    return '#7B388C'
-  }
-  if (route.startsWith('Red') || route.startsWith('Mattapan')) {
-    return '#FA2D27'
-  }
-  if (route.startsWith('Orange')) {
-    return '#FD8A03'
-  }
-  if (
-    route.startsWith('74') ||
-    route.startsWith('75') ||
-    route.startsWith('SL')
-  ) {
-    return '#9A9C9D'
-  }
-  return '#3e2426'
-}
-
-function createVehicleCountMap(): Map<string, Map<string, number>> {
-  const vehicleCountMap = new Map<string, Map<string, number>>()
-  for (const line of lines) {
-    vehicleCountMap.set(line, new Map<string, number>())
-  }
-  return vehicleCountMap
-}
-
-// function clearMap(): void {
-//   for (const line of lines) {
-//     for (const vehicleType of vehicleTypes) {
-//       vehicleCountMap.get(line)?.set(vehicleType, 0)
-//     }
-//   }
-// }
-
-function incrementMapItem(
-  id: string,
-  route: string,
-  vehicleType: string
-): void {
-  // we only want to increment the count if the marker is not already in the map
-  const existingInfo = markerIdInfo.get(id)
-  if (!existingInfo) {
-    markerIdInfo.set(id, { route, vehicleType: vehicleType })
-    const existingCount = vehicleCountMap.get(route)?.get(vehicleType)
-    if (existingCount !== undefined) {
-      vehicleCountMap.get(route)!.set(vehicleType, existingCount + 1)
-    } else {
-      vehicleCountMap.get(route)?.set(vehicleType, 1)
-    }
-  }
-}
-
-function decrementMapItem(id: string): void {
-  const { route, vehicleType } = markerIdInfo.get(id) || {
-    route: '',
-    vehicleType: ''
-  }
-  const existing = vehicleCountMap.get(route)?.get(vehicleType)
-  if (existing !== undefined) {
-    vehicleCountMap.get(route)!.set(vehicleType, existing - 1)
-  }
-}
-
-function calculateTotal(dimension: string): number {
-  let total = 0
-  if (lines.includes(dimension)) {
-    // we are retrieving the total for a line
-    for (const vehicleType of vehicleTypes) {
-      total += vehicleCountMap.get(dimension)?.get(vehicleType) || 0
-    }
-    return total
-  } else if (vehicleTypes.includes(dimension)) {
-    // we are retrieving the total for a vehicle type
-    for (const line of lines) {
-      total += vehicleCountMap.get(line)?.get(dimension) || 0
-    }
-    return total
-  } else if (dimension === 'all') {
-    // we are retrieving the total for all dimensions
-    for (const line of lines) {
-      for (const vehicleType of vehicleTypes) {
-        total += vehicleCountMap.get(line)?.get(vehicleType) || 0
-      }
-    }
-    return total
-  }
-  return total
-}
-
-function updateTable(): void {
-  for (const line of lines) {
-    for (const vehicleType of vehicleTypes) {
-      const id = `${line}-${vehicleType}`
-      const element = document.getElementById(id)
-      if (element) {
-        element.innerHTML = String(
-          DOMPurify.sanitize(
-            String(vehicleCountMap.get(line)?.get(vehicleType) || 0)
-          )
-        )
-      }
-    }
-    const totalElement = document.getElementById(`${line}-total`)
-    if (totalElement) {
-      totalElement.innerHTML = String(
-        DOMPurify.sanitize(String(calculateTotal(line)))
-      )
-    }
-  }
-  for (const vehicleType of vehicleTypes) {
-    const element = document.getElementById(`${vehicleType}-total`)
-    if (element) {
-      element.innerHTML = String(
-        DOMPurify.sanitize(String(calculateTotal(vehicleType)))
-      )
-    }
-  }
-  const element = document.getElementById('total')
-  if (element) {
-    element.innerHTML = String(
-      DOMPurify.sanitize(String(calculateTotal('all')))
-    )
-  }
-}
-
-function getLayerGroupForRoute(route: string): L.LayerGroup {
-  if (route.startsWith('Red') || route.startsWith('Mattapan')) {
-    return layerGroups.red
-  }
-  if (route.startsWith('Blue')) {
-    return layerGroups.blue
-  }
-  if (route.startsWith('Green')) {
-    return layerGroups.green
-  }
-  if (route.startsWith('Orange')) {
-    return layerGroups.orange
-  }
-  if (
-    route.startsWith('SL') ||
-    route.startsWith('74') ||
-    route.startsWith('75')
-  ) {
-    return layerGroups.silver
-  }
-  if (route.startsWith('CR')) {
-    return layerGroups.commuter
-  }
-  if (
-    route === 'Amtrak' ||
-    route.startsWith('Acela') ||
-    route.includes('Northeast Regional') ||
-    route.includes('Downeaster')
-  ) {
-    return layerGroups.amtrak
-  }
-  return layerGroups.commuter
-}
-
-function getShapesLayerGroupForRoute(route: string): L.LayerGroup {
-  if (route.startsWith('Red') || route.startsWith('Mattapan')) {
-    return shapesLayerGroups.red
-  }
-  if (route.startsWith('Blue')) {
-    return shapesLayerGroups.blue
-  }
-  if (route.startsWith('Green')) {
-    return shapesLayerGroups.green
-  }
-  if (route.startsWith('Orange')) {
-    return shapesLayerGroups.orange
-  }
-  if (
-    route.startsWith('SL') ||
-    route.startsWith('74') ||
-    route.startsWith('75')
-  ) {
-    return shapesLayerGroups.silver
-  }
-  if (route.startsWith('CR')) {
-    return shapesLayerGroups.commuter
-  }
-  return shapesLayerGroups.commuter
-}
-
-function pointToLayer(feature: VehicleFeature, latlng: L.LatLng): L.Marker {
-  let icon_size = 28
-  let icon = 'bus-yellow.svg'
-  let opacity = 1.0
-  let zIndex = 0
-  let status = ''
-  let station = ''
-  let stopOrGo = ''
-  // will enable this at a later point once i work out the UI
-  // if (feature.properties["status"] === "STOPPED_AT") {
-  //   stopOrGo = "-stop"
-  // } else if (feature.properties["status"] === "INCOMING_AT" || feature.properties["status"] === "IN_TRANSIT_TO") {
-  //   stopOrGo = "-go"
-  // }
-  if (feature.properties['marker-symbol'] === 'building') {
-    icon = 'entrance-alt1'
-    icon_size = 18
-    opacity = 1
-    zIndex = -10
-  } else {
-    if (feature.properties['marker-symbol'] === 'bus') {
-      opacity = 0.8
-      icon_size = 25
-    }
-    if (feature.properties['marker-size'] === 'small') {
-      icon_size = 27
-    }
-    if (feature.properties['marker-color'] === '#008150') {
-      icon = 'rail-light'
-      incrementMapItem(feature.id as string, 'gl', 'light')
-    }
-    if (feature.properties['marker-color'] === '#2F5DA6') {
-      icon = 'rail-metro-blue'
-      incrementMapItem(feature.id as string, 'bl', 'heavy')
-    }
-    if (feature.properties['marker-color'] === '#FA2D27') {
-      icon = 'rail-metro-red'
-      if (feature.properties['route'] === 'Mattapan') {
-        incrementMapItem(feature.id as string, 'rl', 'light')
-      } else {
-        incrementMapItem(feature.id as string, 'rl', 'heavy')
-      }
-    }
-    if (feature.properties['marker-color'] === '#FD8A03') {
-      icon = 'rail-metro-orange'
-      incrementMapItem(feature.id as string, 'ol', 'heavy')
-    }
-    if (feature.properties['marker-color'] === '#7B388C') {
-      icon = 'rail'
-      incrementMapItem(feature.id as string, 'cr', 'regional')
-    }
-    if (
-      feature.properties.route &&
-      (feature.properties.route === 'Amtrak' ||
-        feature.properties.route.startsWith('Acela') ||
-        feature.properties.route.includes('Northeast Regional'))
-    ) {
-      icon = 'rail-amtrak'
-      incrementMapItem(feature.id as string, 'amtrak', 'regional')
-    }
-    if (feature.properties.route && feature.properties.route.startsWith('SL')) {
-      incrementMapItem(feature.id as string, 'sl', 'bus')
-      icon = 'bus-silver'
-      opacity = 0.9
-    }
-  }
-
-  if (
-    feature.geometry.type === 'Point' &&
-    feature.properties['marker-symbol'] !== 'building'
-  ) {
-    status = feature.properties.status || ''
-    station = feature.properties.stop || ''
-  }
-
-  const leafletIcon = L.icon({
-    iconUrl: `/images/icons/${icon}${stopOrGo}.svg`,
-    iconSize: L.point(icon_size, icon_size)
-  })
-
-  return L.marker(latlng, {
-    icon: leafletIcon,
-    title: `${feature.id} ${feature.properties.route} ${status} ${station}`,
-    opacity: opacity,
-    zIndexOffset: zIndex,
-    riseOnHover: true,
-    riseOffset: 2000
-  })
-}
-
-function niceStatus(status: string): string {
-  if (status === 'INCOMING_AT') {
-    return 'Arriving'
-  }
-  if (status === 'IN_TRANSIT_TO') {
-    return 'In Transit'
-  }
-  if (status === 'STOPPED_AT') {
-    return 'Stopped'
-  }
-  return status
-}
-
-function onEachFeature(feature: VehicleFeature, layer: L.Layer): void {
-  if (feature.geometry.type === 'LineString' && feature.properties.route) {
-    layer.bindPopup(`<b>${feature.properties.route}</b>`)
-  }
-  if (feature.geometry.type === 'Point') {
-    if (feature.properties['marker-symbol'] === 'building') {
-      layer.bindPopup(`<b>${feature.properties.name} Stop</b>`)
-    } else {
-      const update_time = new Date(
-        feature.properties['update_time'] || Date.now()
-      )
-      let speed = ''
-      if (
-        feature.properties['speed'] &&
-        feature.properties['status'] != 'STOPPED_AT'
-      ) {
-        speed = `<br />Speed: ${feature.properties.speed} mph`
-      }
-      if (feature.properties['approximate_speed']) {
-        speed += '* <small>approximate</small>'
-      }
-      let occupancy = ''
-      if (feature.properties['occupancy_status']) {
-        occupancy = `<br />Occupancy: ${feature.properties['occupancy_status']}`
-      }
-      let eta = ''
-      if (feature.properties['stop_eta']) {
-        eta = `<br />ETA: ${feature.properties['stop_eta']}`
-      }
-      let platform_prediction = ''
-      if (
-        !feature.properties.stop?.toLowerCase().includes('track') &&
-        feature.properties['platform_prediction']
-      ) {
-        platform_prediction = `<br />Platform Prediction: ${feature.properties['platform_prediction']}`
-      }
-      const popup = L.popup({
-        content: `<b>${feature.properties.route}/<i>${feature.properties.headsign || feature.properties.stop}</i></b>
-        <br />Stop: ${feature.properties.stop || ''}
-        <br />Status: ${niceStatus(feature.properties.status || '')}
-        ${eta}${speed}${occupancy}${platform_prediction}
-        <br /><small>Update Time: ${update_time.toLocaleTimeString()}</small>`,
-        keepInView: true
-      })
-
-      if (
-        feature.properties['stop-coordinates'] &&
-        feature.properties['status'] != 'STOPPED_AT'
-      ) {
-        const coords: L.LatLngExpression[] = [
-          [feature.geometry.coordinates[1], feature.geometry.coordinates[0]],
-          [
-            feature.properties['stop-coordinates']![1],
-            feature.properties['stop-coordinates']![0]
-          ]
-        ]
-        const line1 = L.polyline(coords, {
-          color: invert(feature.properties['marker-color'] || '#000000', true),
-          weight: 15
-        })
-        const line2 = L.polyline(coords, {
-          color: feature.properties['marker-color'],
-          weight: 7
-        })
-        ;(line1 as any).arrowheads()
-        ;(line2 as any).arrowheads()
-
-        layer.addEventListener('click', () => {
-          line1.addTo(map)
-          line2.addTo(map)
-          map.panInsideBounds(line1.getBounds())
-          window.setTimeout(() => {
-            line1.removeFrom(map)
-            line2.removeFrom(map)
-          }, 10000)
-        })
-      }
-
-      layer.bindPopup(popup)
-    }
-  }
-}
-
-function embedSVG(line: string, alt: string): string {
-  return `<img src="/images/icons/lines/${line}.svg" alt="${alt}" class="line">`
-}
-
-function calculateAffectedLines(data: Array<{ route: string }>): string {
-  const routeMap: Record<string, RouteMapping> = {
-    Red: { svg: 'rl', alt: 'Red Line' },
-    Blue: { svg: 'bl', alt: 'Blue Line' },
-    Greenbush: { svg: 'cr-greenbush', alt: 'Greenbush Line' },
-    Green: { svg: 'gl', alt: 'Green Line' },
-    Orange: { svg: 'ol', alt: 'Orange Line' },
-    749: { svg: 'sl5', alt: 'Silver Line 5' },
-    751: { svg: 'sl4', alt: 'Silver Line 4' },
-    746: { svg: 'slw', alt: 'Silver Line Way' },
-    743: { svg: 'sl3', alt: 'Silver Line 3' },
-    741: { svg: 'sl1', alt: 'Silver Line 1 (Airport)' },
-    742: { svg: 'sl2', alt: 'Silver Line 2' },
-    Fitchburg: { svg: 'cr-fitchburg', alt: 'Fitchburg Line' },
-    Fairmont: { svg: 'cr-fairmont', alt: 'Fairmont Line' },
-    NewBedford: { svg: 'cr-fall-river', alt: 'Fall River/New Bedford Line' },
-    Franklin: { svg: 'cr-franklin', alt: 'Franklin/Foxboro Line' },
-    Haverhill: { svg: 'cr-haverhill', alt: 'Haverhill Line' },
-    Kingston: { svg: 'cr-kingston', alt: 'Kingston Line' },
-    Lowell: { svg: 'cr-lowell', alt: 'Lowell Line' },
-    Needham: { svg: 'cr-needham', alt: 'Needham Line' },
-    Newburyport: { svg: 'cr-newburyport', alt: 'Newburyport/Rockport Line' },
-    Providence: { svg: 'cr-providence', alt: 'Providence Line' },
-    Worcester: { svg: 'cr-worcester', alt: 'Worcester Line' }
-  }
-
-  const afLines = new Set()
-  for (const entity of data) {
-    for (const routePattern in routeMap) {
-      if (
-        entity.route === routePattern ||
-        entity.route.includes(routePattern)
-      ) {
-        const { svg, alt } = routeMap[routePattern]
-        afLines.add(embedSVG(svg, alt))
-        break
-      }
-    }
-  }
-  return [...afLines].join('</br>')
-}
-
-function alerts(): void {
-  $.getJSON(`${vehicles_url}/alerts`, function (data: AlertData) {
-    const msgs = new Set()
-    const dataSet = []
-
-    for (const alert of data.data) {
-      if (alert.attributes && !msgs.has(alert.attributes.header)) {
-        if (
-          alert.attributes.active_period.length > 0 &&
-          alert.attributes.active_period[0].end
-        ) {
-          // skip alert if end time already passed
-          const end_time = alert.attributes.active_period[0].end
-          if (Date.parse(end_time) < Date.now()) {
-            continue
-          }
-        }
-        const rowData = [
-          calculateAffectedLines(alert.attributes.informed_entity),
-          alert.attributes.severity,
-          {
-            display: formatDistance(
-              new Date(
-                alert.attributes.updated_at || alert.attributes.created_at
-              ),
-              new Date(),
-              { addSuffix: true }
-            ),
-            timestamp: new Date(
-              alert.attributes.updated_at || alert.attributes.created_at
-            ).getTime()
-          },
-          alert.attributes.header
-        ]
-        dataSet.push(rowData)
-      }
-    }
-    new DataTable('#alerts', {
-      columns: [
-        { title: 'Lines' },
-        { title: 'Sev', className: 'dt-body-center' },
-        {
-          title: 'Upd',
-          render: {
-            _: 'display',
-            sort: 'timestamp'
-          }
-        },
-        { title: 'Alert', className: 'alert-body' }
-      ],
-      order: [
-        [0, 'desc'],
-        [1, 'desc']
-      ],
-      data: dataSet,
-      ordering: true,
-      paging: false
-    })
-  })
-}
-
-function updateMarkers(features: VehicleFeature[]): void {
-  const newMarkerIds = new Set<string | number>()
-
-  // Process vehicle features (non-building markers)
-  const vehicleFeatures = features.filter(
-    (f) => f.properties['marker-symbol'] !== 'building'
-  )
-
-  for (const feature of vehicleFeatures) {
-    const markerId =
-      feature.id ||
-      `${feature.geometry.coordinates[0]}-${feature.geometry.coordinates[1]}`
-    newMarkerIds.add(markerId)
-
-    const latlng = L.latLng(
-      feature.geometry.coordinates[1],
-      feature.geometry.coordinates[0]
-    )
-    const existingMarker = currentMarkers.get(markerId)
-
-    if (existingMarker) {
-      // Update existing marker position and properties
-      existingMarker.setLatLng(latlng)
-
-      // Update popup content
-      const update_time = new Date(
-        feature.properties['update_time'] || Date.now()
-      )
-      let speed = ''
-      if (
-        feature.properties['speed'] &&
-        feature.properties['status'] != 'STOPPED_AT'
-      ) {
-        speed = `<br />Speed: ${feature.properties.speed} mph`
-      }
-      if (feature.properties['approximate_speed']) {
-        speed += '* <small>approximate</small>'
-      }
-      let occupancy = ''
-      if (feature.properties['occupancy_status']) {
-        occupancy = `<br />Occupancy: ${feature.properties['occupancy_status']}`
-      }
-      let eta = ''
-      if (feature.properties['stop_eta']) {
-        eta = `<br />ETA: ${feature.properties['stop_eta']}`
-      }
-      let platform_prediction = ''
-      if (
-        !feature.properties.stop?.toLowerCase().includes('track') &&
-        feature.properties['platform_prediction']
-      ) {
-        platform_prediction = `<br />Platform Prediction: ${feature.properties['platform_prediction']}`
-      }
-
-      const popupContent = `<b>${feature.properties.route}/<i>${feature.properties.headsign || feature.properties.stop}</i></b>
-        <br />Stop: ${feature.properties.stop || ''}
-        <br />Status: ${niceStatus(feature.properties.status || '')}
-        ${eta}${speed}${occupancy}${platform_prediction}
-        <br /><small>Update Time: ${update_time.toLocaleTimeString()}</small>`
-
-      existingMarker.setPopupContent(popupContent)
-
-      // Update icon if needed
-      const newIcon = createIconForFeature(feature)
-      if (newIcon) {
-        existingMarker.setIcon(newIcon)
-      }
-    } else {
-      // Create new marker
-      const marker = pointToLayer(feature, latlng)
-      currentMarkers.set(markerId, marker)
-
-      if (feature.properties.route) {
-        const layerGroup = getLayerGroupForRoute(feature.properties.route)
-        layerGroup.addLayer(marker)
-      }
-
-      onEachFeature(feature, marker)
-    }
-  }
-
-  // Remove markers that are no longer in the data
-  for (const [markerId, marker] of currentMarkers.entries()) {
-    if (!newMarkerIds.has(markerId)) {
-      // Remove from all layer groups
-      Object.values(layerGroups).forEach((group) => {
-        if (group.hasLayer(marker)) {
-          group.removeLayer(marker)
-        }
-      })
-      currentMarkers.delete(markerId)
-      decrementMapItem(markerId as string)
-    }
-  }
-}
-
-function createIconForFeature(feature: VehicleFeature): L.Icon | null {
-  let icon_size = 28
-  let icon = 'bus-yellow.svg'
-  let stopOrGo = ''
-
-  if (feature.properties['marker-symbol'] === 'building') {
-    return null // Buildings handled separately
-  }
-
-  if (feature.properties['marker-symbol'] === 'bus') {
-    icon_size = 25
-  }
-  if (feature.properties['marker-size'] === 'small') {
-    icon_size = 27
-  }
-  if (feature.properties['marker-color'] === '#008150') {
-    icon = 'rail-light'
-  }
-  if (feature.properties['marker-color'] === '#2F5DA6') {
-    icon = 'rail-metro-blue'
-  }
-  if (feature.properties['marker-color'] === '#FA2D27') {
-    icon = 'rail-metro-red'
-  }
-  if (feature.properties['marker-color'] === '#FD8A03') {
-    icon = 'rail-metro-orange'
-  }
-  if (feature.properties['marker-color'] === '#7B388C') {
-    icon = 'rail'
-  }
-  if (
-    feature.properties.route &&
-    (feature.properties.route === 'Amtrak' ||
-      feature.properties.route.startsWith('Acela') ||
-      feature.properties.route.includes('Northeast Regional'))
-  ) {
-    icon = 'rail-amtrak'
-  }
-  if (feature.properties.route && feature.properties.route.startsWith('SL')) {
-    icon = 'bus-silver'
-  }
-
-  return L.icon({
-    iconUrl: `/images/icons/${icon}${stopOrGo}.svg`,
-    iconSize: L.point(icon_size, icon_size)
-  })
-}
 
 function annotate_map(): void {
   $.getJSON(vehicles_url, function (data: any) {
@@ -800,7 +74,7 @@ function annotate_map(): void {
   })
 
   // Fetch Amtrak data from BOS API
-  fetchAmtrakData()
+  fetchAmtrakData(bos_url)
   if (!baseLayerLoaded) {
     Object.values(shapesLayerGroups).forEach((group) => {
       group.clearLayers()
@@ -837,7 +111,7 @@ function annotate_map(): void {
             )
             shapesGroup.addLayer(layer)
           }
-          onEachFeature(feature as VehicleFeature, layer)
+          onEachFeature(feature as any, layer)
         }
       })
     })
@@ -894,7 +168,6 @@ L.easyButton({
 // Variables to store the user location circle and watch state
 let userLocationCircle: L.Circle | null = null
 let userLocationMarker: L.Marker | null = null
-// let isWatchingLocation: boolean = false
 let locationWatchID: number | null = null
 
 // Handle successful geolocation
@@ -915,19 +188,6 @@ map.on('locationfound', (e: L.LocationEvent) => {
     fillOpacity: 0.2,
     weight: 2
   }).addTo(map)
-
-  // Add a marker at the exact location
-  // userLocationMarker = L.marker(e.latlng, {
-  //   icon: L.icon({
-  //     iconUrl: '/images/icons/location-dot.svg',
-  //     iconSize: L.point(20, 20)
-  //   })
-  // }).addTo(map).bindPopup('Your location')
-
-  // // If not watching location, center the map view once
-  // if (!isWatchingLocation) {
-  //   map.setView(e.latlng, Math.max(map.getZoom(), 15))
-  // }
 })
 
 // Handle geolocation errors
@@ -1037,7 +297,6 @@ document
 // Function to start/stop location watching
 function toggleLocationWatch(enabled: boolean): void {
   if (enabled) {
-    // isWatchingLocation = true
     locationWatchID = navigator.geolocation.watchPosition(
       (position) => {
         const lat = position.coords.latitude
@@ -1065,7 +324,6 @@ function toggleLocationWatch(enabled: boolean): void {
       }
     )
   } else {
-    // isWatchingLocation = false
     if (locationWatchID !== null) {
       navigator.geolocation.clearWatch(locationWatchID)
       locationWatchID = null
@@ -1095,41 +353,4 @@ document
     setCookie('follow-location', target.checked.toString())
   })
 
-function fetchAmtrakData(): void {
-  $.getJSON(`${bos_url}/trains/geojson`, function (data: any) {
-    if (data && data.features) {
-      L.geoJSON(data, {
-        pointToLayer: (feature: any, latlng: L.LatLng) => {
-          const transformedFeature: VehicleFeature = {
-            id: feature.id,
-            geometry: feature.geometry,
-            properties: {
-              route: 'Amtrak',
-              'marker-color': '#004080',
-              'marker-symbol': 'rail',
-              status: feature.properties.status || 'IN_TRANSIT_TO',
-              headsign:
-                feature.properties.headsign ||
-                feature.properties.route ||
-                'Amtrak',
-              speed: feature.properties.speed || 0,
-              update_time:
-                (feature.properties as any).timestamp ||
-                new Date().toISOString()
-            }
-          }
-
-          const marker = pointToLayer(transformedFeature, latlng)
-          layerGroups.amtrak.addLayer(marker)
-          return marker
-        },
-        onEachFeature: onEachFeature as any,
-        filter: (feature) => {
-          return feature.geometry.type === 'Point'
-        }
-      })
-    }
-  })
-}
-
-alerts()
+alerts(vehicles_url)
